fix: Merge utils.py into utils/__init__.py to resolve module shadowing
- Moved all functions from utils.py into utils/__init__.py
- Deleted utils.py file that was shadowing the utils/ package directory
- This allows both PDF extractor imports and utility function imports to work
- Resolves ModuleNotFoundError for sk_agents.utils.pdf_extractor
