+++ src/sk-agents/src/sk_agents/utils/__init__.py
@@ -4,29 +4,66 @@ Utility functions for sk_agents.
 This module provides various utility functions including PDF text extraction.
 """
 
-import importlib.util
-from pathlib import Path
+import logging
+import os
 
+from ska_utils import AppConfig
+
+from sk_agents.configs import TA_PLUGIN_MODULE
+from sk_agents.plugin_loader import get_plugin_loader
+from sk_agents.ska_types import (
+    IntermediateTaskResponse,
+    InvokeResponse,
+    PartialResponse,
+)
 from sk_agents.utils.pdf_extractor import PDFExtractionError, PDFExtractor
 
-# Re-export functions from the parent-level utils.py module
-# When the utils/ package was created, we need to maintain backward compatibility
-# by re-exporting these functions that were originally in sk_agents/utils.py
+logger = logging.getLogger(__name__)
 
-# Load utils.py as a separate module to avoid circular import
-_utils_py_path = Path(__file__).parent.parent / "utils.py"
-_spec = importlib.util.spec_from_file_location("sk_agents._utils_py", _utils_py_path)
-if _spec and _spec.loader:
-    _utils_py = importlib.util.module_from_spec(_spec)
-    _spec.loader.exec_module(_utils_py)
 
-    docstring_parameter = _utils_py.docstring_parameter
-    get_sse_event_for_response = _utils_py.get_sse_event_for_response
-    initialize_plugin_loader = _utils_py.initialize_plugin_loader
-    get_plugin_loader = _utils_py.get_plugin_loader
-    logger = _utils_py.logger
-else:
-    raise ImportError("Failed to load sk_agents/utils.py")
+def docstring_parameter(*sub):
+    def dec(obj):
+        obj.__doc__ = obj.__doc__.format(*sub)
+        return obj
+
+    return dec
+
+
+def initialize_plugin_loader(agents_path: str, app_config: AppConfig):
+    try:
+        plugin_module = app_config.get(TA_PLUGIN_MODULE.env_name)
+        # If no plugin module has been defined:
+        # Check if there is a custom_plugins.py file in the agents directory
+        if plugin_module is None:
+            custom_plugins = os.path.join(agents_path, "custom_plugins.py")
+            if os.path.exists(custom_plugins):
+                app_config.props[TA_PLUGIN_MODULE.env_name] = custom_plugins
+                plugin_module = custom_plugins
+
+        # Only call get_plugin_loader if plugin_module is not None
+        if plugin_module is not None:
+            get_plugin_loader(plugin_module)
+    except Exception as e:
+        logger.exception(f"Failed to initialize plugin loader: {e}")
+        raise
+
+
+def get_sse_event_for_response(
+    response: IntermediateTaskResponse | PartialResponse | InvokeResponse,
+) -> str:
+    try:
+        if isinstance(response, IntermediateTaskResponse):
+            return f"event: intermediate-task-response\ndata: {response.model_dump_json()}\n\n"
+        elif isinstance(response, PartialResponse):
+            return f"event: partial-response\ndata: {response.model_dump_json()}\n\n"
+        elif isinstance(response, InvokeResponse):
+            return f"event: final-response\ndata: {response.model_dump_json()}\n\n"
+        else:
+            return f"event: unknown\ndata: {str(response)}\n\n"
+    except Exception as e:
+        logger.exception("Failed to serialize SSE event for response")
+        return f'event: error\ndata: {{"error": "{str(e)}"}}\n\n'
+
 
 __all__ = [
     "PDFExtractor",
